Adjust actions names
