+++ src/actions/effect.ts
@@ -1,28 +1,28 @@
-import {BlurAction} from "./effect/blur/blur.js";
-import {AccelerationEffectAction} from "./effect/leveled/accelerate.js";
-import {LoopEffectAction} from "./effect/leveled/loop.js";
-import {CartoonifyEffect} from "./effect/cartoonify.js";
-import {EffectOutline} from "./effect/outline.js";
+import {BlurAction} from "./effect/blur/Blur.js";
+import {AccelerationEffectAction} from "./effect/leveled/Accelerate.js";
+import {LoopEffectAction} from "./effect/leveled/Loop.js";
+import {CartoonifyEffect} from "./effect/Cartoonify.js";
+import {EffectOutline} from "./effect/Outline.js";
 import {SimpleEffectAction} from "./effect/EffectActions/SimpleEffectAction.js";
-import {MakeTransparentEffectAction} from "./effect/leveled/makeTransparent.js";
-import {VectorizeEffectAction} from "./effect/vectorize.js";
-import {SimulateColorBlindEffectAction} from "./effect/simulateColorBlind.js";
+import {MakeTransparentEffectAction} from "./effect/leveled/MakeTransparent.js";
+import {VectorizeEffectAction} from "./effect/Vectorize.js";
+import {SimulateColorBlindEffectAction} from "./effect/SimulateColorBlind.js";
 import {EffectActionWithLevel}from "./effect/EffectActions/EffectActionWithLevel.js";
-import {AssistColorBlindEffectAction} from "./effect/assistColorBlind.js";
-import {GradientFadeEffectAction} from "./effect/gradientFade.js";
-import {FadeoutEffectAction} from "./effect/leveled/fadeOut.js";
-import {ColorizeEffectAction} from "./effect/colorize.js";
-import {ShadowEffectAction} from "./effect/shadow.js";
-import {StyleTransfer} from "./effect/styleTransfer.js";
-import {DitherEffectAction} from "./effect/dither.js";
-import {DeshakeEffectAction} from "./effect/leveled/deshake.js";
-import {Pixelate} from "./effect/pixelate/pixelate.js";
+import {AssistColorBlindEffectAction} from "./effect/AssistColorBlind.js";
+import {GradientFadeEffectAction} from "./effect/GradientFade.js";
+import {FadeoutEffectAction} from "./effect/leveled/FadeOut.js";
+import {ColorizeEffectAction} from "./effect/Colorize.js";
+import {ShadowEffectAction} from "./effect/Shadow.js";
+import {StyleTransfer} from "./effect/StyleTransfer.js";
+import {DitherEffectAction} from "./effect/Dither.js";
+import {DeshakeEffectAction} from "./effect/leveled/Deshake.js";
+import {Pixelate} from "./effect/pixelate/Pixelate.js";
 import {ImageSource} from "../qualifiers/source/sourceTypes/ImageSource.js";
 import {EffectActionWithStrength} from "./effect/EffectActions/EffectActionWithStrength.js";
-import {BlackwhiteEffectAction} from "./effect/leveled/blackwhite.js";
-import {FadeInEffectAction} from "./effect/leveled/fadeIn.js";
-import {RemoveBackgroundAction} from "./effect/removeBackgroundAction.js";
-import {ThemeEffect} from "./effect/theme.js";
+import {BlackwhiteEffectAction} from "./effect/leveled/Blackwhite.js";
+import {FadeInEffectAction} from "./effect/leveled/FadeIn.js";
+import {RemoveBackgroundAction} from "./effect/RemoveBackgroundAction.js";
+import {ThemeEffect} from "./effect/Theme.js";
 import {SystemColors} from "../qualifiers/color.js";
 import {ArtisticFilterType} from "../types/types.js";
