+++ src/actions/effect/Dither.ts
