+++ src/actions/effect/GradientFade.ts
