+++ src/actions/effect/RemoveBackgroundAction.ts
