+++ src/actions/effect/SimulateColorBlind.ts
