+++ src/actions/effect/Vectorize.ts
