+++ src/actions/effect/leveled/FadeOut.ts
