+++ src/actions/effect/pixelate/Pixelate.ts
