--- src/actions/roundCorners/max.ts
@@ -1,14 +1,0 @@
-import RoundCornersAction from "./RoundCornersAction.js";
-
-
-/**
- * @summary action
- * @description Generates an asset with a circular crop using the 'max' radius value.
- * @memberOf Actions.RoundCorners
- * @return {Actions.RoundCorners.RoundCornersAction}
- */
-function max():RoundCornersAction {
-  return new RoundCornersAction().max();
-}
-
-export default max;
